- removed lazy instantiation - moved class up a level from their interfaces - extracted more descriptive methods - removed final keyword - boiler plate code - more fluent method calls where possible - added my email address as bug report email address. - compiled using new version of intellij (ipr-file has been changed)
