+++ src/org/intellij/ideaplugins/tabswitch/action/NextTabAction.java
@@ -22,13 +22,10 @@ import com.intellij.openapi.vfs.VirtualFile;
 
 public class NextTabAction extends TabAction {
 
-  private FileFetcher<VirtualFile> fileFetcher;
+  private final FileFetcher<VirtualFile> fileFetcher = new FileFetcherOpenTabFiles();
 
   @Override
   protected FileFetcher<VirtualFile> getFileFetcher() {
-    if (fileFetcher == null) {
-      fileFetcher = new FileFetcherOpenTabFiles();
-    }
     return fileFetcher;
   }
