+++ src/org/intellij/ideaplugins/tabswitch/errorReporter.properties
@@ -1,1 +1,1 @@
-email.to=basleijdekkers@gmail.com
+email.to=must.tahir@gmail.com
